Refactoring drupal8 naming to drupal
