+++ tests/Fixtures/ConfigResolver/ComposerDrupal/composer.json
